refactor: make NondeterministicLRUCacheEvictor extends LRU
