+++ core/client/fs/src/main/java/alluxio/client/file/cache/evictor/LRUCacheEvictor.java
@@ -32,7 +32,7 @@ public class LRUCacheEvictor implements CacheEvictor {
   private static final boolean UNUSED_MAP_VALUE = true;
 
   // TODO(feng): unify with worker side evictor
-  private final Map<PageId, Boolean> mLRUCache =
+  protected final Map<PageId, Boolean> mLRUCache =
       Collections.synchronizedMap(new LinkedHashMap<>(LINKED_HASH_MAP_INIT_CAPACITY,
           LINKED_HASH_MAP_INIT_LOAD_FACTOR, LINKED_HASH_MAP_ACCESS_ORDERED));
